editing the rateing,donations,slider

+++ mainproject/models.py
@@ -20,10 +20,9 @@ class Catogrey(models.Model):
 
 class Project(models.Model):
 
-    
     title=models.CharField(max_length=100)
     details=models.CharField(max_length=100)
-    total_target=models.DecimalField(max_digits=10,decimal_places=2)
+    total_target=models.FloatField()
     creator=models.ForeignKey(settings.AUTH_USER_MODEL , on_delete=models.CASCADE, blank=True, null=True)
     start_time=models.DateTimeField()
     end_time=models.DateTimeField()
@@ -53,7 +52,7 @@ class Comment(models.Model):
 
 
 class Donation(models.Model):
-    donation = models.DecimalField(max_digits=10,decimal_places=2)
+    donation = models.FloatField()
     project = models.ForeignKey(Project, on_delete=models.CASCADE)
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
